fix: Calls and missed calls notifications should be ignored when the user is Away (SQCORE-1180)
https://wearezeta.atlassian.net/browse/SQCORE-1180

Currently we have an inconsistent behaviour when a call comes in while the availability status of the user is Away.
The call is ignored, but the "missed call" notification appears.

This PR fixes it by ensuring that any call notification will be ignored when the user is Away, including missed calls.