feat: Assets in Federation (FS-293) (#3643)
https://wearezeta.atlassian.net/browse/FS-293

Implements downloading assets through the v4 endpoints and handling the information about the domain
where the backend should search for the given asset. Works for files, pictures, profile pictures,
audio and video. (Locations still don't display previews).

The PR also fixes unit tests which failed before when the federation flag was on.

There is one significant change to how the user search works which applies also to non-federation
environments. Until now the results of a search were not stored in the database. It was actually
a change from an earlier implementation where we did that, and with more complicated handling of
profile pictures I decided to go back to this idea. I think it shouldn't make the search more heavy,
or heavier that we can tolerate. I will test it some more.