refactor: Remove old code dealing with v2 assets (FS-293) (#3639)
* refactor: Remove old code dealing with v2 assets (FS-293)

https://wearezeta.atlassian.net/browse/FS-293

As a preparation for work on federated assets, this PR removes the old code for v2 assets.

* feat: Federation: Changes to protobuf, assets and RemoteData (FS-158) (#3640)

https://wearezeta.atlassian.net/browse/FS-158

We have a new generic-message-proto that adds domains fields to Assets to RemoteData.
This PR updates Android's generic-message-proto to this new version, makes changes to Scala classes
handling these fields, and introduces a new migration in the database.
Also, I removed some dead code.

* feat: Assets in Federation (FS-293) (#3643)

https://wearezeta.atlassian.net/browse/FS-293

Implements downloading assets through the v4 endpoints and handling the information about the domain
where the backend should search for the given asset. Works for files, pictures, profile pictures,
audio and video. (Locations still don't display previews).

The PR also fixes unit tests which failed before when the federation flag was on.

There is one significant change to how the user search works which applies also to non-federation
environments. Until now the results of a search were not stored in the database. It was actually
a change from an earlier implementation where we did that, and with more complicated handling of
profile pictures I decided to go back to this idea. I think it shouldn't make the search more heavy,
or heavier that we can tolerate. I will test it some more.