fix: Make profile pictures more responsive to changes (SQCORE-1227)
https://wearezeta.atlassian.net/browse/SQCORE-1227

My recent work on assets in Federation exposed a small issue in how profile pictures are being updated.
Sometimes when another user changes their profile picture, or even when the current user requests
the profile picture of another user, the new picture is not displayed in all places, for example
in little icons next to messages.

This consists of two changes
1. Every "chathead" listens now to update events.
2. I added headers for "content type" and "accept" to a request for the profile picture
(without the headers it is possible that the backend will refuse).