fix: Repeated missed calls notifications and calling screens (SQCORE-1133)
https://wearezeta.atlassian.net/browse/SQCORE-1133

If the app is in the background, a call is received and dismissed, and then the app is opened, the calling screen appears again.
It happens because in the new implementation it is possible for the app to process the same event twice - first for the notification
and then when the app is opened. In case of ther types of events this is desirable, in the case of call events it is not.
This PR fixes that by remembering the time of the last call event and expecting new events to be later that this time.

Also, missed calls notifications were ignoring rules for when the notifications might appear. I refactored their logic a bit
to make sure it's not the case anymore.