feat: Ask the backend for the list of classified domains (FS-385)
https://wearezeta.atlassian.net/browse/FS-385

If the app is connected to a federated backend, and the federation flag is on, on start the app will now ask
the backend for the list of classified domains. The response will be stored in the user properties and used
later (not in this commit) for displaying a banner on the conversation view if the conversation is classified
and safe. A failure to obtain the response, or in general on non-federated backends, we will have an empty list
of classified domains, but also a boolean flag telling us that no banner should be displayed - the feature is
disabled.