fix(build): remove 'null' from F-Droid build version strings. (#3662)
The version_name_postfix field is not set in the build configuration for
F-Droid, which means that an erroneous "null" is appended to the version
strings. For example, the F-Droid build of client version 3.78.13 contains
the version string 3.78.13null in the APK metadata. This breaks the F-Droid
project's automated builds of the application, as the version string embedded
in the APK does not match the version given in the build instructions.

This commit removes the version name postfix from the F-Droid build
configuration, as this functionality is unused.

Co-authored-by: Marco Bierbach <Gardosen@users.noreply.github.com>