fix: Limit notifications reappering on the screen after they are swiped away (SQCORE-1133) (#3667)
https://wearezeta.atlassian.net/browse/SQCORE-1133

I suspect that there might be a bug in Android, resulting in that a swiped away notification is not really deleted and can come back. I want to limit the number of times we ask Android to display new notifications and this way don't trigger it into displaying the old ones together with the new ones. I created a registry for unique ids of displayed notifications (a unique id consists of the account to which it belongs to, the conversation it belongs to, and the timestamp when it arrived from the backend). When a notification is going to be displayed, just before that we check if we haven't displayed it already, and if yes, we don't do it again.

This is basically a hack, but it looks like it limits the number of those annoying recurrences.