fix: The "reply" notification should be used only for replies to the current user (SQCORE-1249)
https://wearezeta.atlassian.net/browse/SQCORE-1249

Currently this form of notification is used for any reply, also when another participant of a group conversation
replies to yet another participant.
To check if the reply is the reply to the current user we need to look for the original message (the one that is
replied to) and then check if the author of that message is the current participant.