feat: Securpty clearance on the call view (FS-385)
https://wearezeta.atlassian.net/browse/FS-385

This PR adds the security clearance banner on the call view, as designed in the JIRA ticket.
The banner responds to the securpty clearance of participants of the conversation, not those only in the call.
The banner changes in real time.

The calling functionality is separated from the text-based conversations functionality so I had to copy-paste some of the code.
I would prefer to leave it like this instead of making some complex connections between classes only to avoid duplication.