feat: Use feature config to enable/disable guest links (SQSERVICES-1050) (#3670)
https://wearezeta.atlassian.net/browse/SQSERVICES-1050

There is a new feature flag on the backend telling us if we should allow guest links, i.e. both creating guest rooms
and inviting guests with links, and join conversations as guests using those links.
By default the flag is "enabled" - meaning that if the backend does not give us any answer we will allow guest links,
and so nothing will change for a user who updates to Wire Android with this code, unless the backend explicitly
states otherwise.