feat: Show the security clearance banner for unconnected users (FS-385)
https://wearezeta.atlassian.net/browse/FS-385

The information about the security clearance for unconnected users, i.e. on connect request view, pending connection view, etc.,
need to be computed and displayed differently from the standard conversatino participant view. It required some refactoring
to be able to use the same code to check for clearance in both cases, and also I changed the name of one case object from
`ClassifiedConnection.NotConnected` to `ClassifiedConnection.Unconnected` to be consistent with the new name for that case,
hence so many changes.