fix: The guest room button in SearchUI (SQSERVICES-1050)
https://wearezeta.atlassian.net/browse/SQSERVICES-1050

The button for creating a guest room, visible in the SearchUI, now depends on the guest links feature flag fetched from the backend.
The bug caused the button to disappear even if the flag was enabled. This is because in RetrieveSearchResults, where we configure
the SearchUI, it's not allowed to use Scala futures - all the code needs to be synchronous - so we need to check the flag by
`.currentValue` which gives us `Option[Boolean]` instead of `.head` which gives us `Future[Boolean]`.

Another problem was that I made a horrible copy-paste error in `FeatureConfigsClient` and the feature flag wasn't actually fetched at all.