fix: The pending connect request view was blank (FS-385) (#3676)
https://wearezeta.atlassian.net/browse/FS-385

In the new implementation of the view the user id of the other user is kept in a signal (previously it was a constant).
A signal needed initialization before it can be used and the pending connection request view didn't do it,
because it overwrote the method that was doing the initialization without calling `super` (i.e. the original one).

I added the call to the original initialization method and all looks good.