fix(search): remove external team members from team searches (FS-506) (#3684)
When the backend does not report the domain for a user (maybe they don't support federation yet), the code checked that the user is on a different domain than you and assumed that was not an external team member.

I changed the code by removing the part that checks for domain. If a user has the same team ID as you, and their permission schema matches that of an external user, then they are external. In the rare case where there is a team ID collision across backends, this means that the user is indeed from another team and not yours, and in this case their permission schema is not available to you, therefore the conditions will (correctly) evaluate to false.