fix: Make the event processing recover from failures
The bug that makes some of the messages never appear in the conversations is not 100% reproducible,
but its behaviour seems consistent with a flaw in our event processing - if an event can't be decrypted,
or if it's decrypted but then its processing fails, it will be deleted from the storage anyway,
even though it's not displayed. On top of that, the complexity of how we process events hides from us
some of the errors.

In this commit I bypass the event processing algorithm, create another, much simpler one, and thanks to that
I'm able to delete from the storage only those events that were decrypted and processed successfully.
For events that failed to be processed, the processing is retried two times, and only after the third time
the app assumes nothing can be done and delete them. But in the process we get much more explicit error
messages which may help us identifying underlying bugs which result in failed processing.

This change is experimental. I didn't remove the dead code of the old algorithm. Also, even if the new version
works well, some easy optimizations can be done.

Please create an Internal build and let people test it. This way we may either identify underlying bugs,
fix them, and then go back to the old algorithm, or replace the old algorithm with this one, but make it
a bit better still.