One fix and more log statements
I found something weird in Event.scala in how qualified ids for users and conversations are serialized/deserialized.
I fixed it but I'm not sure if this is connected to the bug. I looked into how conversation data is searched for
using remote ids and now I believe the search is alright - it does not use any cache that could be not up to date
when called for the first time. Instead, please look at lines 506 and 583 in Event.scala. It looks like in some situations
Wire Android serialized an already deserialized event back to json, and then deserializes it again. I suspect
this might be messed up.

I added additional log statements to Event.scala. Please test again.