fix: use android-job version published in Sonatype instead of local fork.
The version of android-job which we depend on has not been updated for
AndroidX, and building from a bundled copy in this repository does not allow us
to run Jetifier to make it compatible. We've published our fork on Sonatype so
that we can pull it in via the usual third-party dependency mechanisms which go
through Jetifier.